# Improve interface type membership lookup