Looking over some profiling artifacts, I noticed this call to `interface_type_memberships` wasn't cached, so objects were repeatedly building this array over and over. The more interfaces and superclasses an object has, the more this overhead would've added up. I was able to confirm a reduction in allocated memory and time spent calculating this by running the new benchmark: 

```diff 
# ... 
-        176   (0.0%)          25   (0.0%)     GraphQL::Schema::Member::HasFields#get_field
+          4   (0.0%)           1   (0.0%)     GraphQL::Schema::Member::HasFields#get_field

-         32   (0.0%)          17   (0.0%)     GraphQL::Schema::Member::HasInterfaces#interface_type_memberships
+          4   (0.0%)           3   (0.0%)     GraphQL::Schema::Warden#interface_type_memberships


# ... 
- Total allocated: 12937088 bytes (69487 objects)
+ Total allocated: 12935984 bytes (69454 objects)

# ...
-         40  /Users/rmosolgo/code/graphql-ruby/lib/graphql/schema/member/has_interfaces.rb
-         39  /Users/rmosolgo/code/graphql-ruby/lib/graphql/schema/member/has_fields.rb
-         37  /Users/rmosolgo/code/graphql-ruby/lib/graphql/schema/warden.rb
+         40  /Users/rmosolgo/code/graphql-ruby/lib/graphql/schema/warden.rb
+         39  /Users/rmosolgo/code/graphql-ruby/lib/graphql/schema/member/has_fields.rb
# ...
+          4  /Users/rmosolgo/code/graphql-ruby/lib/graphql/schema/member/has_interfaces.rb
```

TODO: 

- [ ] Forward-port to 2.0.x