@rmosolgo 2023-02-14
This doesn't seem to have made a difference in real life, the benchmark even shows it 0.03% _slower_ 🤷 


  &nbsp; | Total | Self | Method 
  ---|---|---|---
  **Before** | 18.02ms (0.09%) | 0.00ns (<0.01%) | GraphQL::Schema::Member::HasFields#visible_interface_implementation?
**After** | 25.00ms (0.12%) | 6.24ms (0.03%) | GraphQL::Schema::Member::HasFields#visible_interface_implementation?